notebbok, help and view
